requirements.txt
customtkinter==5.2.1
pillow==10.2.0
psutil==5.9.8
pywin32